global-news-tracker/requirements.txt
feedparser

newspaper3k
lxml_html_clean           # Required for newspaper3k to work

#  Summarization using Hugging Face
transformers
torch


streamlit


# Networking & Parsing (used internally or for safety)
requests